+++ commands/version/lib/git-tag.js
@@ -5,10 +5,14 @@ const childProcess = require("@lerna/child-process");
 
 module.exports = gitTag;
 
-function gitTag(tag, { signGitTag }, opts) {
+function gitTag(tag, { forceGitTag, signGitTag }, opts) {
   log.silly("gitTag", tag);
 
   const args = ["tag", tag, "-m", tag];
+
+  if (forceGitTag) {
+    args.push("--force");
+  }
 
   if (signGitTag) {
     args.push("--sign");